sample_app/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0
